+++ src/command_modules/azure-cli-advisor/azure/cli/command_modules/advisor/__init__.py
@@ -3,12 +3,12 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import azure.cli.command_modules.component._help  # pylint: disable=unused-import
+import azure.cli.command_modules.advisor._help  # pylint: disable=unused-import
 
 
 def load_params(_):
-    import azure.cli.command_modules.component._params  # pylint: disable=redefined-outer-name, unused-variable
+    import azure.cli.command_modules.advisor._params    # pylint: disable=redefined-outer-name, unused-variable
 
 
 def load_commands():
-    import azure.cli.command_modules.component.commands  # pylint: disable=redefined-outer-name, unused-variable
+    import azure.cli.command_modules.advisor.commands   # pylint: disable=redefined-outer-name, unused-variable