+++ src/azure-cli/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.0.22
+++++++
+* Remove `az component` commands. Use `az extension` instead. `az component` has been deprecated for several months now.
+
 2.0.21
 ++++++
 * Minor fixes